+++ src/Sulu/Bundle/Sales/OrderBundle/DependencyInjection/SuluSalesOrderExtension.php
@@ -10,7 +10,7 @@
 
 namespace Sulu\Bundle\Sales\OrderBundle\DependencyInjection;
 
-use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Sulu\Bundle\PersistenceBundle\DependencyInjection\PersistenceExtensionTrait;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
@@ -18,6 +18,8 @@ use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
 class SuluSalesOrderExtension extends Extension
 {
+    use PersistenceExtensionTrait;
+
     /**
      * {@inheritDoc}
      */
@@ -80,6 +82,8 @@ class SuluSalesOrderExtension extends Extension
             'sulu_sales_order.pdf_order_dynamically_name_prefix',
             $config['pdf_order_dynamically_name_prefix']
         );
+
+        $this->configurePersistence($config['objects'], $container);
     }
 
     /**